Broken stuff :-(

+++ vagrant/startup.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+for i in $*; do
+  echo $i
+done
+
 count=$(docker ps -a -q | wc -l)
 if [ "$count" -gt 0 ]; then
     echo "Stopping and removing all running containers"
@@ -14,19 +18,21 @@ docker run -d -e LOGSTASH_CONFIG_URL=https://raw.githubusercontent.com/witlox/cu
 docker pull djbnjack/mongobase
 docker run -d --net=host --name="mongodb" djbnjack/mongobase
 
-docker pull witlox/cumulonimbi
-docker run -d --net=host --name="jobmanager" witlox/cumulonimbi python cumulonimbi/cumulonimbi.py -a
+if [ "$1" != "local-only" ]; then
+    docker pull witlox/cumulonimbi
+    docker run -d --net=host --name="jobmanager" witlox/cumulonimbi python cumulonimbi/cumulonimbi.py -a
 
-echo "Waiting for logstash and the api to start."
-until $(curl --output /dev/null --silent --head --fail http://localhost:9200); do
-    printf '.'
-    sleep 10
-done
-until $(curl --output /dev/null --silent --head --fail http://localhost:5000/jobs); do
-    printf '.'
-    sleep 10
-done
-echo "All up!"
+    echo "Waiting for logstash and the api to start."
+    until $(curl --output /dev/null --silent --head --fail http://localhost:9200); do
+        printf '.'
+        sleep 10
+    done
+    until $(curl --output /dev/null --silent --head --fail http://localhost:5000/jobs); do
+        printf '.'
+        sleep 10
+    done
+    echo "All up!"
 
-echo "Running integration tests"
-docker run -i --rm --net=host --name="integrationtests" witlox/cumulonimbi nosetests cumulonimbi/tests/integrationtests
+    echo "Running integration tests"
+    docker run -i --rm --net=host --name="integrationtests" witlox/cumulonimbi nosetests cumulonimbi/tests/integrationtests
+fi
